+++ seleniumtid/visual_test.py
@@ -13,6 +13,7 @@ been supplied.
 
 import logging
 import os
+import shutil
 
 from seleniumtid import selenium_driver
 
@@ -26,6 +27,11 @@ except ImportError:
 
 
 class VisualTest(object):
+    template_name = 'VisualTestsTemplate.html'
+    row_template_name = 'VisualTestsRowTemplate.html'
+    css_name = 'bootstrap.min.css'
+    report_name = 'VisualTests.html'
+
     def __init__(self):
         if not selenium_driver.config.getboolean_optional('Server', 'visualtests_enabled'):
             return
@@ -45,6 +51,7 @@ class VisualTest(object):
         if not self.save_baseline:
             if not os.path.exists(self.output_directory):
                 os.makedirs(self.output_directory)
+                self.copy_template()
 
     def assertScreenshot(self, element_or_selector, filename, file_suffix, threshold=0):
         """Assert that a screenshot of an element is the same as a screenshot on disk, within a given threshold.
@@ -87,7 +94,49 @@ class VisualTest(object):
             try:
                 self.engine.assertSameFiles(output_file, baseline_file, threshold)
             except AssertionError as exc:
+                self.add_to_report(file_suffix, output_file, baseline_file)
                 if selenium_driver.config.getboolean_optional('Server', 'visualtests_fail'):
                     raise exc
                 else:
                     self.logger.warn('Visual error: {}'.format(exc.message))
+
+    def copy_template(self):
+        """Copy html template and css file to output directory"""
+        template_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'resources', self.template_name)
+        shutil.copyfile(template_path, os.path.join(self.output_directory, self.report_name))
+        css_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'resources', self.css_name)
+        shutil.copyfile(css_path, os.path.join(self.output_directory, self.css_name))
+
+    def get_html_row(self, test_name, output_file, baseline_file):
+        """Create the html row with the result of a visual test
+
+        :param test_name: test name
+        :param output_file: output file path
+        :param baseline_file: baseline file path
+        :returns: str with the html row
+        """
+        img = '<img style="width: 100%" onclick="window.open(this.src)" src="file://{}"/></td>'
+        row = '<tr>'
+        row += '<td>' + test_name + '</td>'
+        row += '<td>' + img.format(baseline_file) + '</td>'
+        row += '<td>' + img.format(output_file) + '</td>'
+        diff_file = output_file.replace('.png', '.diff.png')
+        diff_row = img.format(diff_file) if os.path.exists(diff_file) else ''
+        row += '<td>' + diff_row + '</td>'
+        row += '</tr>'
+        return row
+
+    def add_to_report(self, test_name, output_file, baseline_file):
+        """Add the result of a visual test to the html report
+
+        :param test_name: test name
+        :param output_file: output file path
+        :param baseline_file: baseline file path
+        """
+        row = self.get_html_row(test_name, output_file, baseline_file)
+        with open(os.path.join(self.output_directory, self.report_name), "r+") as f:
+            report = f.read()
+            index = report.find('</tbody>')
+            report = report[:index] + row + report[index:]
+            f.seek(0)
+            f.write(report)